requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
openai>=1.0.0
PyPDF2>=3.0.0
flask>=3.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
selenium>=4.16.0
webdriver-manager>=4.0.2

